client/shared/koji_utils.py: add koji_utils back as an autotest library
As evidenced by the kernelinstall test, this library can be used by tests
that are not necessarily related to virtualization. So, add it back to
the autotest client library.

This version has no difference in functionality from the version found
in the next branch of the virt-test repo as of this date. It has the
following cosmetics differences though:

 * docstrings in sphinx format
 * fixes apointed by pylint (mostly variable naming)
 * pylint disable statements for false positives

Signed-off-by: Cleber Rosa <crosa@redhat.com>